refactor(editor): Repoint imports at the relocated modal components (no-changelog)
Relative imports that pointed into the shell become absolute `@/app/...`; the
shell's `Modals.vue` now imports each modal from its owning feature. No
registration change — the keys stay where they are.
