+++ packages/frontend/editor-ui/src/app/components/Modals.vue
@@ -70,7 +70,7 @@ import {
 import { API_KEY_CREATE_OR_EDIT_MODAL_KEY } from '@/features/settings/apiKeys/apiKeys.constants';
 import type { ApiKeyWithRawValue } from '@n8n/api-types';
 import AboutModal from '@/app/components/AboutModal.vue';
-import ActivationModal from '@/app/components/ActivationModal.vue';
+import ActivationModal from '@/features/workflows/components/ActivationModal.vue';
 import ApiKeyCreateOrEditModal from '@/features/settings/apiKeys/components/ApiKeyCreateOrEditModal.vue';
 import NewAssistantSessionModal from '@/features/ai/assistant/components/Chat/NewAssistantSessionModal.vue';
 import ChatEmbedModal from '@/app/components/ChatEmbedModal.vue';
@@ -83,14 +83,14 @@ import DebugPaywallModal from '@/features/execution/executions/components/DebugP
 import DeleteFolderModal from '@/features/core/folders/components/DeleteFolderModal.vue';
 import MoveToFolderModal from '@/features/core/folders/components/MoveToFolderModal.vue';
 import DeleteUserModal from '@/features/settings/users/components/DeleteUserModal.vue';
-import DuplicateWorkflowDialog from '@/app/components/DuplicateWorkflowDialog.vue';
+import DuplicateWorkflowDialog from '@/features/workflows/components/DuplicateWorkflowDialog.vue';
 import ExternalSecretsProviderModal from '@/features/integrations/externalSecrets.ee/components/ExternalSecretsProviderModal.ee.vue';
 import SecretsProviderConnectionModal from '@/features/integrations/secretsProviders.ee/components/SecretsProviderConnectionModal.ee.vue';
 import DeleteSecretsProviderModal from '@/features/integrations/secretsProviders.ee/components/DeleteSecretsProviderModal.ee.vue';
-import FromAiParametersModal from '@/app/components/FromAiParametersModal.vue';
+import FromAiParametersModal from '@/features/ndv/parameters/components/FromAiParametersModal.vue';
 import ImportCurlModal from '@/features/ndv/parameters/components/ImportCurlModal.vue';
 import BinaryDataViewModal from '@/features/ndv/runData/components/BinaryDataViewModal.vue';
-import ImportWorkflowUrlModal from '@/app/components/ImportWorkflowUrlModal.vue';
+import ImportWorkflowUrlModal from '@/features/workflows/components/ImportWorkflowUrlModal.vue';
 import InviteUsersModal from '@/features/settings/users/components/InviteUsersModal.vue';
 import ModalRoot from '@/app/components/ModalRoot.vue';
 import NpsSurvey from '@/app/components/NpsSurvey.vue';
@@ -104,27 +104,27 @@ import SourceControlPushModal from '@/features/integrations/sourceControl.ee/com
 import AnnotationTagsManager from '@/features/shared/tags/components/TagsManager/AnnotationTagsManager.ee.vue';
 import WorkflowTagsManager from '@/features/shared/tags/components/TagsManager/WorkflowTagsManager.vue';
 import WhatsNewModal from '@/app/components/WhatsNewModal.vue';
-import WorkflowActivationConflictingWebhookModal from '@/app/components/WorkflowActivationConflictingWebhookModal.vue';
-import WorkflowExtractionNameModal from '@/app/components/WorkflowExtractionNameModal.vue';
+import WorkflowActivationConflictingWebhookModal from '@/features/workflows/components/WorkflowActivationConflictingWebhookModal.vue';
+import WorkflowExtractionNameModal from '@/features/workflows/components/WorkflowExtractionNameModal.vue';
 import WorkflowHistoryVersionUnpublishModal from '@/features/workflows/workflowHistory/components/WorkflowHistoryVersionUnpublishModal.vue';
 import WorkflowVersionFormModal, {
 	type WorkflowVersionFormModalData,
 } from '@/features/workflows/workflowHistory/components/WorkflowVersionFormModal.vue';
-import WorkflowSettings from '@/app/components/WorkflowSettings/WorkflowSettings.vue';
-import WorkflowShareModal from '@/app/components/WorkflowShareModal.ee.vue';
+import WorkflowSettings from '@/features/workflows/components/WorkflowSettings/WorkflowSettings.vue';
+import WorkflowShareModal from '@/features/workflows/components/WorkflowShareModal.ee.vue';
 import WorkflowDiffModal from '@/features/workflows/workflowDiff/WorkflowDiffModal.vue';
 import type { EventBus } from '@n8n/utils/event-bus';
 import DynamicModalLoader from './DynamicModalLoader.vue';
 import NodeRecommendationModalV2 from '@/experiments/templateRecoV2/components/NodeRecommendationModal.vue';
 import NodeRecommendationModalV3 from '@/experiments/personalizedTemplatesV3/components/NodeRecommendationModal.vue';
 import VariableModal from '@/features/settings/environments.ee/components/VariableModal.vue';
-import StopManyExecutionsModal from './StopManyExecutionsModal.vue';
+import StopManyExecutionsModal from '@/features/execution/executions/components/StopManyExecutionsModal.vue';
 import AddExecutionToDatasetModal from '@/features/ai/evaluation.ee/components/AddExecutionToDataset/AddExecutionToDatasetModal.vue';
-import WorkflowDescriptionModal from '@/app/components/WorkflowDescriptionModal.vue';
-import WorkflowPublishModal from '@/app/components/MainHeader/WorkflowPublishModal.vue';
+import WorkflowDescriptionModal from '@/features/workflows/components/WorkflowDescriptionModal.vue';
+import WorkflowPublishModal from '@/features/workflows/components/WorkflowPublishModal.vue';
 import MigrateWorkflowModal from '@/features/settings/migrationReport/MigrateWorkflowModal.vue';
 import UpdatesPanel from './UpdatesPanel.vue';
-import CredentialResolverEditModal from '@/app/components/CredentialResolverEditModal.vue';
+import CredentialResolverEditModal from '@/features/resolvers/components/CredentialResolverEditModal.vue';
 import AIBuilderDiffModal from '@/features/ai/assistant/components/Agent/AIBuilderDiffModal.vue';
 import AiGatewayTopUpModal from '@/features/ai/gateway/components/AiGatewayTopUpModal.vue';
 import { defineAsyncComponent } from 'vue';
