+++ packages/frontend/editor-ui/src/features/resolvers/components/CredentialResolverEditModal.test.ts
@@ -1,10 +1,10 @@
 import { createComponentRenderer } from '@/__tests__/render';
 import { type MockedStore, mockedStore } from '@/__tests__/utils';
 import { createTestingPinia } from '@pinia/testing';
-import CredentialResolverEditModal from '@/app/components/CredentialResolverEditModal.vue';
+import CredentialResolverEditModal from '@/features/resolvers/components/CredentialResolverEditModal.vue';
 import { useRootStore } from '@n8n/stores/useRootStore';
 import { useToast } from '@n8n/composables/useToast';
-import { CREDENTIAL_RESOLVER_EDIT_MODAL_KEY } from '../constants';
+import { CREDENTIAL_RESOLVER_EDIT_MODAL_KEY } from '@/app/constants';
 import * as restApiClient from '@n8n/rest-api-client';
 import type { CredentialResolverType } from '@n8n/api-types';
 import { defineComponent, h, watch, toRefs } from 'vue';
