+++ packages/frontend/editor-ui/src/features/workflows/components/WorkflowDescriptionModal.vue
@@ -12,9 +12,9 @@ import {
 } from '@/app/stores/workflowDocument.store';
 import { useToast } from '@n8n/composables/useToast';
 import { useTelemetry } from '@n8n/composables/useTelemetry';
-import { WORKFLOW_DESCRIPTION_MODAL_KEY } from '../constants';
+import { WORKFLOW_DESCRIPTION_MODAL_KEY } from '@/app/constants';
 import { createEventBus } from '@n8n/utils/event-bus';
-import Modal from './Modal.vue';
+import Modal from '@/app/components/Modal.vue';
 import WorkflowTagsDropdown from '@/features/shared/tags/components/WorkflowTagsDropdown.vue';
 import { onMounted } from 'vue';
