+++ packages/frontend/editor-ui/src/features/workflows/components/WorkflowPublishModal.test.ts
@@ -1,7 +1,7 @@
 import { createTestingPinia } from '@pinia/testing';
 import { createComponentRenderer } from '@/__tests__/render';
 import { type MockedStore, mockedStore } from '@/__tests__/utils';
-import WorkflowPublishModal from '@/app/components/MainHeader/WorkflowPublishModal.vue';
+import WorkflowPublishModal from '@/features/workflows/components/WorkflowPublishModal.vue';
 import { useWorkflowsStore } from '@/app/stores/workflows.store';
 import { useWorkflowsListStore } from '@/app/stores/workflowsList.store';
 import { useNodeTypesStore } from '@/app/stores/nodeTypes.store';
